requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pypdf
python-dotenv
pdfminer.six